Move print to bin/vcloud-walker
This makes walk method testable. Also adds a test for it. It's not a great test because the data is mocked, but it acts as a check that the correct methods are being called.

The error message when not one of the valid calls has been removed. Will be returned in a later commit.