use fog 1.21
This allows the user to stop holding credentials in a .fog file.

This includes the raw fog changes only. To use it you'll need to get a
session token, by externally logging in, and set that as FOG_VCLOUD_TOKEN
ENV var.
see https://github.com/fog/fog/pull/2705 for further detail.