Prevent token from being logged by Jenkins
Disable `set -x` in all Jenkins scripts so that the FOG_VCLOUD_TOKEN
environment variable is never logged by Jenkins. These scripts are fairly
simple so we shouldn't lose much by disabling bash debugging.

And explicitly disable debugging with `set +x` just before we set the token,
to guard against the possibility of someone unwittingly re-enabling this in
the future.